requirements.txt
python-telegram-bot~=20.7
asyncio>=3.4.3
python_version=">=3.8"
dotenv